@@ -1,37 +1,40 @@
 {
-  "name": "rbexample",
-  "version": "0.0.1",
-  "main": "node_modules/expo/AppEntry.js",
-  "private": true,
-  "scripts": {
-    "start": "react-native-scripts start",
-    "eject": "react-native-scripts eject",
-    "android": "react-native-scripts android",
-    "ios": "react-native-scripts ios",
-    "postinstall": "rm -rf node_modules/native-base-shoutem-theme/node_modules/hoist-non-react-statics node_modules/xdl/node_modules/react node_modules/react-google-maps/node_modules/prop-types/"
-  },
-  "dependencies": {
-    "@babel/polyfill": "^7.0.0-rc.3",
-    "@ptomasroos/react-native-multi-slider": "^0.0.13",
-    "appbase-js": "^4.0.2-beta.05",
-    "escape-string-regexp": "^1.0.5",
-    "expo": "^27.0.1",
-    "hoist-non-react-statics": "^2.5.0",
-    "metro": "^0.34.0",
-    "metro-core": "^0.34.0",
-    "native-base": "^2.4.2",
-    "react": "16.3.0-alpha.1",
-    "react-native": "^0.55.4",
-    "react-native-calendars": "^1.18.2",
-    "react-redux": "^5.0.7",
-    "redux": "3.7.2",
-    "redux-thunk": "^2.2.0",
-    "xdate": "^0.8.2"
-  },
-  "devDependencies": {
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-preset-expo": "^4.0.0",
-    "prop-types": "^15.6.1",
-    "react-native-scripts": "^1.14.0"
-  }
+	"name": "rbexample",
+	"version": "0.0.1",
+	"main": "node_modules/expo/AppEntry.js",
+	"private": true,
+	"scripts": {
+		"start": "EXPO_DEBUG=true expo start",
+		"eject": "expo eject",
+		"android": "expo start --android",
+		"ios": "expo start --ios",
+		"postinstall": "rm -rf node_modules/native-base-shoutem-theme/node_modules/hoist-non-react-statics node_modules/xdl/node_modules/react node_modules/react-google-maps/node_modules/prop-types/ node_modules/react-native-gesture-handler/node_modules/hoist-non-react-statics"
+	},
+	"dependencies": {
+		"@babel/polyfill": "^7.0.0-rc.3",
+		"@babel/runtime": "7.4.3",
+		"@ptomasroos/react-native-multi-slider": "^1.0.0",
+		"appbase-js": "^4.0.2-beta.05",
+		"escape-string-regexp": "^1.0.5",
+		"expo": "^32.0.6",
+		"hoist-non-react-statics": "^3.3.0",
+		"native-base": "^2.4.2",
+		"react": "16.8.6",
+		"react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+		"react-native-calendars": "^1.18.2",
+		"react-redux": "^5.0.7",
+		"redux": "4.0.1",
+		"redux-thunk": "^2.2.0",
+		"xdate": "^0.8.2"
+	},
+	"devDependencies": {
+		"babel-plugin-module-resolver": "^3.2.0",
+		"babel-preset-expo": "^5.1.1",
+		"prop-types": "^15.7.2",
+		"react-native-scripts": "^2.0.1",
+		"schedule": "0.4.0"
+	},
+	"resolutions": {
+		"@babel/runtime": "7.4.3"
+	}
 }
