@@ -1,6 +1,6 @@
 {
 	"name": "@appbaseio/reactivesearch-native",
-	"version": "0.11.1",
+	"version": "0.12.0",
 	"description": "Native data components for building reactive UIs with Elasticsearch",
 	"main": "src/index.js",
 	"files": [
@@ -40,15 +40,15 @@
 		"eslint-plugin-react-native": "^3.2.1",
 		"husky": "^0.14.3",
 		"prettier": "^1.12.1",
-		"react": "~16.3.2",
-		"react-native": "~0.55.4"
+		"react": "~16.8.6",
+		"react-native": "~0.59.1"
 	},
 	"dependencies": {
-		"@appbaseio/reactivecore": "^8.1.0",
+		"@appbaseio/reactivecore": "^8.1.2",
 		"@babel/polyfill": "^7.0.0-rc.3",
-		"@ptomasroos/react-native-multi-slider": "^0.0.13",
+		"@ptomasroos/react-native-multi-slider": "^1.0.0",
 		"appbase-js": "^4.0.2-beta.05",
-		"hoist-non-react-statics": "^2.5.0",
+		"hoist-non-react-statics": "^3.3.0",
 		"native-base": "^2.4.2",
 		"prop-types": "^15.6.1",
 		"react-native-calendars": "^1.18.2",
