@@ -34,13 +34,6 @@ class GeoDistanceDropdown extends Component {
 	constructor(props) {
 		super(props);
 
-		this.state = {
-			currentLocation: null,
-			currentDistance: 0,
-			userLocation: null,
-			suggestions: [],
-			isOpen: false,
-		};
 		this.type = 'geo_distance';
 		this.locked = false;
 		this.coordinates = null;
@@ -49,74 +42,98 @@ class GeoDistanceDropdown extends Component {
 		if (props.autoLocation) {
 			this.getUserLocation();
 		}
+
+		let currentLocation = null;
+		let currentDistance = 0;
+
 		props.setQueryListener(props.componentId, props.onQueryChange, null);
-	}
+		props.addComponent(this.props.componentId);
+		this.setReact(props);
 
-	componentWillMount() {
-		this.props.addComponent(this.props.componentId);
-		this.setReact(this.props);
+		if (props.value) {
+			currentLocation = props.value.location;
+			const selected = props.data.find(
+				item => item.label === props.value.label,
+			);
 
-		if (this.props.selectedValue) {
-			this.setState({
-				currentLocation: this.props.selectedValue.location,
-			});
-			this.getCoordinates(this.props.selectedValue.location, () => {
-				const selected = this.props.data.find(
-					item => item.label === this.props.selectedValue.label,
-				);
-				this.setDistance(selected.distance);
-			});
-		} else if (this.props.defaultSelected) {
-			this.setState({
-				currentLocation: this.props.defaultSelected.location,
-			});
-			this.getCoordinates(this.props.defaultSelected.location, () => {
-				const selected = this.props.data.find(
-					item => item.label === this.props.defaultSelected.label,
-				);
-				this.setDistance(selected.distance);
-			});
+			currentDistance = selected.distance;
+		} else if (props.selectedValue) {
+			currentLocation = props.selectedValue.location;
+			const selected = props.data.find(
+				item => item.label === props.selectedValue.label,
+			);
+
+			currentDistance = selected.distance;
+		} else if (props.defaultValue) {
+			currentLocation = props.defaultValue.location;
+			const selected = props.data.find(
+				item => item.label === props.defaultValue.label,
+			);
+			currentDistance = selected.distance;
 		}
+
+
+		this.state = {
+			currentLocation,
+			currentDistance,
+			userLocation: null,
+			suggestions: [],
+			isOpen: false,
+		};
+		this.getCoordinates(currentLocation, () => {
+			this.setDistance(currentDistance);
+		});
 	}
 
 	componentDidMount() {
 		this.autocompleteService = new window.google.maps.places.AutocompleteService();
 	}
 
-	componentWillReceiveProps(nextProps) {
-		checkPropChange(this.props.react, nextProps.react, () => this.setReact(nextProps));
+	componentDidUpdate(prevProps) {
+		checkPropChange(this.props.react, prevProps.react, () => this.setReact(this.props));
 
-		checkSomePropChange(this.props, nextProps, ['dataField', 'nestedField'], () => {
-			this.updateQuery(this.state.currentDistance, nextProps);
+		checkSomePropChange(this.props, prevProps, ['dataField', 'nestedField'], () => {
+			this.updateQuery(this.state.currentDistance, this.props);
 		});
 
 		if (
-			nextProps.defaultSelected
-			&& nextProps.defaultSelected.label
-			&& nextProps.defaultSelected.location
-			&& !isEqual(this.props.defaultSelected, nextProps.defaultSelected)
+			this.props.value
+			&& !isEqual(this.props.value, prevProps.value)
 		) {
-			this.setValues(nextProps.defaultSelected, nextProps);
+			this.setValues(this.props.value, this.props);
 		} else if (
-			nextProps.selectedValue
-			&& nextProps.selectedValue.label
-			&& nextProps.selectedValue.location
-			&& !isEqual(this.state.currentLocation, nextProps.selectedValue.location)
+			this.props.selectedValue
+			&& this.props.selectedValue.label
+			&& this.props.selectedValue.location
+			&& !isEqual(this.state.currentLocation, this.props.selectedValue.location)
+			&& !isEqual(this.props.selectedValue, prevProps.selectedValue)
 		) {
-			this.setValues(nextProps.selectedValue, nextProps);
+			const { value, onChange } = this.props;
+
+			if (value === undefined) {
+				this.setValues(this.props.selectedValue, this.props);
+			} else if (onChange) {
+				onChange(this.props.selectedValue);
+			}
 		} else if (
-			!isEqual(this.props.selectedValue, nextProps.selectedValue)
-			&& !nextProps.selectedValue
+			!isEqual(this.props.selectedValue, prevProps.selectedValue)
+			&& !this.props.selectedValue
 		) {
-			this.setState(
-				{
-					currentLocation: null,
-					currentDistance: null,
-				},
-				() => {
-					this.updateQuery(null);
-				},
-			);
+			const { value, onChange } = this.props;
+			if (value === undefined) {
+				// eslint-disable-next-line
+				this.setState(
+					{
+						currentLocation: null,
+						currentDistance: null,
+					},
+					() => {
+						this.updateQuery(null);
+					},
+				);
+			} else if (onChange) {
+				onChange(null);
+			}
 		}
 	}
 
@@ -131,11 +148,12 @@ class GeoDistanceDropdown extends Component {
 	}
 
 	setValues = (selected, props) => {
+		const selectedDistance = props.data.find(item => item.label === selected.label);
 		this.setState({
 			currentLocation: selected.location,
+			currentDistance: selectedDistance.distance,
 		});
 		this.getCoordinates(selected.location, () => {
-			const selectedDistance = props.data.find(item => item.label === selected.label);
 			this.setDistance(selectedDistance.distance);
 		});
 	};
@@ -295,14 +313,28 @@ class GeoDistanceDropdown extends Component {
 	};
 
 	onDistanceChange = (value) => {
-		this.setDistance(value.distance);
+		const { onChange, value: valueProp } = this.props;
+		if (valueProp === undefined) {
+			this.setDistance(value.distance);
+		} else if (onChange) {
+			onChange({ label: value.label, location: this.state.currentLocation });
+		}
 	};
 
 	onInputChange = (e) => {
 		const { value } = e.target;
-		this.setState({
-			currentLocation: value,
-		});
+		const { onChange, value: propValue } = this.props;
+
+		if (propValue === undefined) {
+			this.setState({
+				currentLocation: value,
+			});
+		} else if (onChange) {
+			onChange({
+				location: value,
+				label: this.props.value.label,
+			});
+		}
 		if (value.trim()) {
 			if (!this.autocompleteService) {
 				this.autocompleteService = new window.google.maps.places.AutocompleteService();
@@ -346,7 +378,16 @@ class GeoDistanceDropdown extends Component {
 	};
 
 	handleOuterClick = () => {
-		this.setLocation({ value: this.state.currentLocation });
+		const { onChange, value } = this.props;
+
+		if (value === undefined) {
+			this.setLocation({ value: this.state.currentLocation });
+		} else if (onChange) {
+			onChange({
+				location: this.state.currentLocation,
+				label: this.props.value.label,
+			});
+		}
 	};
 
 	handleStateChange = (changes) => {
@@ -357,6 +398,19 @@ class GeoDistanceDropdown extends Component {
 			});
 		}
 	};
+
+	handleLocation = (data) => {
+		const { value, onChange } = this.props;
+
+		if (value === undefined) {
+			this.setLocation(data);
+		} else if (onChange) {
+			onChange({
+				location: data.value,
+				label: this.props.value.label,
+			});
+		}
+	}
 
 	renderSearchBox = () => {
 		let suggestionsList = [...this.state.suggestions];
@@ -374,7 +428,7 @@ class GeoDistanceDropdown extends Component {
 
 		return (
 			<Downshift
-				onChange={this.setLocation}
+				onChange={this.handleLocation}
 				onOuterClick={this.handleOuterClick}
 				onStateChange={this.handleStateChange}
 				isOpen={this.state.isOpen}
@@ -481,7 +535,7 @@ GeoDistanceDropdown.propTypes = {
 	customQuery: types.func,
 	data: types.data,
 	dataField: types.stringRequired,
-	defaultSelected: types.selectedValue,
+	defaultValue: types.selectedValue,
 	filterLabel: types.string,
 	icon: types.children,
 	iconPosition: types.iconPosition,
@@ -489,6 +543,7 @@ GeoDistanceDropdown.propTypes = {
 	innerRef: types.func,
 	nestedField: types.string,
 	onBlur: types.func,
+	onChange: types.func,
 	onFocus: types.func,
 	onKeyDown: types.func,
 	onKeyPress: types.func,
@@ -497,6 +552,7 @@ GeoDistanceDropdown.propTypes = {
 	onValueChange: types.func,
 	placeholder: types.string,
 	react: types.react,
+	value: types.selectedValue,
 	showFilter: types.bool,
 	showIcon: types.bool,
 	style: types.style,
