@@ -1,7 +1,13 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 
-import { ReactiveBase, DataSearch, ResultList, SelectedFilters } from '@appbaseio/reactivesearch';
+import {
+	ReactiveBase,
+	DataSearch,
+	ResultList,
+	ReactiveList,
+	SelectedFilters,
+} from '@appbaseio/reactivesearch';
 
 import './index.css';
 
@@ -44,56 +50,76 @@ class Main extends Component {
 								return filters;
 							}}
 						/>
-						<ResultList
+						<ReactiveList
 							componentId="SearchResult"
 							dataField="original_title"
 							from={0}
 							size={3}
-							renderItem={this.booksList}
 							className="result-list-container"
 							pagination
 							react={{
 								and: 'BookSensor',
 							}}
+							render={({ data }) => (
+								<ReactiveList.ResultListWrapper>
+									{data.map(item => (
+										<ResultList key={item._id}>
+											<ResultList.Image src={item.image} />
+											<ResultList.Content>
+												<ResultList.Title>
+													<div
+														className="book-title"
+														dangerouslySetInnerHTML={{
+															__html: item.original_title,
+														}}
+													/>
+												</ResultList.Title>
+												<ResultList.Description>
+													<div className="flex column justify-space-between">
+														<div>
+															<div>
+																by{' '}
+																<span className="authors-list">
+																	{item.authors}
+																</span>
+															</div>
+															<div className="ratings-list flex align-center">
+																<span className="stars">
+																	{Array(
+																		item.average_rating_rounded,
+																	)
+																		.fill('x')
+																		.map((item, index) => (
+																			<i
+																				className="fas fa-star"
+																				key={index}
+																			/>
+																		)) // eslint-disable-line
+																	}
+																</span>
+																<span className="avg-rating">
+																	({item.average_rating} avg)
+																</span>
+															</div>
+														</div>
+														<span className="pub-year">
+															Pub {item.original_publication_year}
+														</span>
+													</div>
+												</ResultList.Description>
+											</ResultList.Content>
+										</ResultList>
+									))}
+								</ReactiveList.ResultListWrapper>
+							)}
 						/>
 					</div>
 				</div>
 			</ReactiveBase>
 		);
 	}
-
-	booksList(data) {
-		return {
-			title: (
-				<div
-					className="book-title"
-					dangerouslySetInnerHTML={{ __html: data.original_title }}
-				/>
-			),
-			description: (
-				<div className="flex column justify-space-between">
-					<div>
-						<div>
-							by <span className="authors-list">{data.authors}</span>
-						</div>
-						<div className="ratings-list flex align-center">
-							<span className="stars">
-								{Array(data.average_rating_rounded)
-									.fill('x')
-									.map((item, index) => (
-										<i className="fas fa-star" key={index} />
-									)) // eslint-disable-line
-								}
-							</span>
-							<span className="avg-rating">({data.average_rating} avg)</span>
-						</div>
-					</div>
-					<span className="pub-year">Pub {data.original_publication_year}</span>
-				</div>
-			),
-			image: data.image,
-		};
-	}
 }
+
+export default Main;
 
 ReactDOM.render(<Main />, document.getElementById('root'));
