@@ -1,79 +1,95 @@
-import React, { Component } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { ReactiveBase, DataSearch, ResultList, SelectedFilters } from '@appbaseio/reactivesearch';
+import {
+	ReactiveBase,
+	DataSearch,
+	ReactiveList,
+	ResultCard,
+	SelectedFilters,
+} from '@appbaseio/reactivesearch';
 
 import './index.css';
 
-class Main extends Component {
-	render() {
-		return (
-			<ReactiveBase
-				app="good-books-ds"
-				credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d"
-			>
-				<div className="row">
-					<div className="col">
-						<DataSearch
-							title="DataSearch"
-							dataField={['original_title', 'original_title.search']}
-							categoryField="authors.raw"
-							componentId="BookSensor"
-							URLParams
-						/>
-					</div>
-
-					<div className="col">
-						<SelectedFilters />
-						<ResultList
-							componentId="SearchResult"
-							dataField="original_title"
-							size={10}
-							renderItem={this.renderData}
-							className="result-list-container"
-							pagination
-							react={{
-								and: 'BookSensor',
-							}}
-						/>
-					</div>
-				</div>
-			</ReactiveBase>
-		);
-	}
-
-	renderData(data) {
-		return {
-			title: (
-				<div
-					className="book-title"
-					dangerouslySetInnerHTML={{ __html: data.original_title }}
+const Main = () => (
+	<ReactiveBase app="good-books-ds" credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d">
+		<div className="row">
+			<div className="col">
+				<DataSearch
+					title="DataSearch"
+					dataField={['original_title', 'original_title.search']}
+					categoryField="authors.raw"
+					componentId="BookSensor"
+					URLParams
 				/>
-			),
-			description: (
-				<div className="flex column justify-space-between">
-					<div>
-						<div>
-							by <span className="authors-list">{data.authors}</span>
-						</div>
-						<div className="ratings-list flex align-center">
-							<span className="stars">
-								{Array(data.average_rating_rounded)
-									.fill('x')
-									.map((item, index) => (
-										<i className="fas fa-star" key={index} />
-									)) // eslint-disable-line
-								}
-							</span>
-							<span className="avg-rating">({data.average_rating} avg)</span>
-						</div>
-					</div>
-					<span className="pub-year">Pub {data.original_publication_year}</span>
-				</div>
-			),
-			image: data.image,
-		};
-	}
-}
+			</div>
+
+			<div className="col">
+				<SelectedFilters />
+				<ReactiveList
+					componentId="SearchResult"
+					dataField="original_title"
+					size={10}
+					className="result-list-container"
+					pagination
+					react={{
+						and: 'BookSensor',
+					}}
+					render={({ data }) => (
+						<ReactiveList.ResultCardsWrapper>
+							{data.map(item => (
+								<ResultCard key={item.id}>
+									<ResultCard.Image src={item.image} />
+									<ResultCard.Title>
+										<div
+											className="book-title"
+											dangerouslySetInnerHTML={{
+												__html: item.original_title,
+											}}
+										/>
+									</ResultCard.Title>
+
+									<ResultCard.Description>
+										<div className="flex column justify-space-between">
+											<div>
+												<div>
+													by{' '}
+													<span className="authors-list">
+														{item.authors}
+													</span>
+												</div>
+												<div className="ratings-list flex align-center">
+													<span className="stars">
+														{Array(item.average_rating_rounded)
+															.fill('x')
+															.map((item, index) => (
+																<i
+																	className="fas fa-star"
+																	key={index}
+																/>
+															)) // eslint-disable-line
+														}
+													</span>
+													<span className="avg-rating">
+														({item.average_rating} avg)
+													</span>
+												</div>
+											</div>
+											<span className="pub-year">
+												Pub {item.original_publication_year}
+											</span>
+										</div>
+									</ResultCard.Description>
+								</ResultCard>
+							))}
+						</ReactiveList.ResultCardsWrapper>
+					)}
+				/>
+			</div>
+		</div>
+	</ReactiveBase>
+);
+
+export default Main;
 
 ReactDOM.render(<Main />, document.getElementById('root'));
