@@ -2,7 +2,13 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import moment from 'moment';
 
-import { ReactiveBase, DatePicker, ResultCard, SelectedFilters } from '@appbaseio/reactivesearch';
+import {
+	ReactiveBase,
+	DatePicker,
+	ResultCard,
+	SelectedFilters,
+	ReactiveList,
+} from '@appbaseio/reactivesearch';
 
 import './index.css';
 
@@ -26,16 +32,49 @@ class Main extends Component {
 
 					<div className="col">
 						<SelectedFilters />
-						<ResultCard
+						<ReactiveList
 							componentId="SearchResult"
 							dataField="name"
 							from={0}
 							size={40}
-							renderItem={this.renderData}
 							showPagination
 							react={{
 								and: ['DateSensor'],
 							}}
+							render={({ data }) => (
+								<ReactiveList.ResultCardsWrapper>
+									{data.map(item => (
+										<ResultCard href={item.listing_url} key={item.id}>
+											<ResultCard.Image src={item.image} />
+											<ResultCard.Title>
+												<div
+													className="book-title"
+													dangerouslySetInnerHTML={{
+														__html: item.name,
+													}}
+												/>
+											</ResultCard.Title>
+
+											<ResultCard.Description>
+												<div>
+													<div>${item.price}</div>
+													<span
+														style={{
+															backgroundImage: `url(${
+																item.host_image
+															})`,
+														}}
+													/>
+													<p>
+														{item.room_type} · {item.accommodates}{' '}
+														guests
+													</p>
+												</div>
+											</ResultCard.Description>
+										</ResultCard>
+									))}
+								</ReactiveList.ResultCardsWrapper>
+							)}
 						/>
 					</div>
 				</div>
@@ -57,23 +96,6 @@ class Main extends Component {
 			];
 		}
 		return query ? { query: { bool: { must: query } } } : null;
-	}
-
-	renderData(res) {
-		return {
-			image: res.image,
-			title: res.name,
-			description: (
-				<div>
-					<div>${res.price}</div>
-					<span style={{ backgroundImage: `url(${res.host_image})` }} />
-					<p>
-						{res.room_type} · {res.accommodates} guests
-					</p>
-				</div>
-			),
-			url: res.listing_url,
-		};
 	}
 }
 
