@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
 import {
@@ -6,81 +6,93 @@ import {
 	DynamicRangeSlider,
 	SelectedFilters,
 	ResultList,
+	ReactiveList,
 } from '@appbaseio/reactivesearch';
 
 import './index.css';
 
-export default class Main extends Component {
-	render() {
-		return (
-			<ReactiveBase
-				app="good-books-ds"
-				credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d"
-			>
-				<div className="row">
-					<div className="col">
-						<DynamicRangeSlider
-							dataField="books_count"
-							componentId="BookSensor"
-							rangeLabels={(min, max) => ({
-								start: `${min} book`,
-								end: `${max} books`,
-							})}
-						/>
-					</div>
-
-					<div className="col">
-						<SelectedFilters />
-						<ResultList
-							componentId="SearchResult"
-							dataField="original_title"
-							from={0}
-							size={3}
-							className="result-list-container"
-							renderItem={this.booksList}
-							pagination
-							react={{
-								and: 'BookSensor',
-							}}
-						/>
-					</div>
-				</div>
-			</ReactiveBase>
-		);
-	}
-
-	booksList(data) {
-		return {
-			title: (
-				<div
-					className="book-title"
-					dangerouslySetInnerHTML={{ __html: data.original_title }}
+const Main = () => (
+	<ReactiveBase app="good-books-ds" credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d">
+		<div className="row">
+			<div className="col">
+				<DynamicRangeSlider
+					dataField="books_count"
+					componentId="BookSensor"
+					rangeLabels={(min, max) => ({
+						start: `${min} book`,
+						end: `${max} books`,
+					})}
 				/>
-			),
-			description: (
-				<div className="flex column justify-space-between">
-					<div>
-						<div>
-							by <span className="authors-list">{data.authors}</span>
-						</div>
-						<div className="ratings-list flex align-center">
-							<span className="stars">
-								{Array(data.average_rating_rounded)
-									.fill('x')
-									.map((item, index) => (
-										<i className="fas fa-star" key={index} />
-									)) // eslint-disable-line
-								}
-							</span>
-							<span className="avg-rating">({data.average_rating} avg)</span>
-						</div>
-					</div>
-					<span className="pub-year">Pub {data.original_publication_year}</span>
-				</div>
-			),
-			image: data.image,
-		};
-	}
-}
+			</div>
+
+			<div className="col">
+				<SelectedFilters />
+				<ReactiveList
+					componentId="SearchResult"
+					dataField="original_title"
+					from={0}
+					size={3}
+					className="result-list-container"
+					pagination
+					react={{
+						and: 'BookSensor',
+					}}
+					render={({ data }) => (
+						<ReactiveList.ResultListWrapper>
+							{data.map(item => (
+								<ResultList key={item._id}>
+									<ResultList.Image src={item.image} />
+									<ResultList.Content>
+										<ResultList.Title>
+											<div
+												className="book-title"
+												dangerouslySetInnerHTML={{
+													__html: item.original_title,
+												}}
+											/>
+										</ResultList.Title>
+										<ResultList.Description>
+											<div className="flex column justify-space-between">
+												<div>
+													<div>
+														by{' '}
+														<span className="authors-list">
+															{item.authors}
+														</span>
+													</div>
+													<div className="ratings-list flex align-center">
+														<span className="stars">
+															{Array(item.average_rating_rounded)
+																.fill('x')
+																.map((item, index) => (
+																	<i
+																		className="fas fa-star"
+																		key={index}
+																	/>
+																)) // eslint-disable-line
+															}
+														</span>
+														<span className="avg-rating">
+															({item.average_rating} avg)
+														</span>
+													</div>
+												</div>
+												<span className="pub-year">
+													Pub {item.original_publication_year}
+												</span>
+											</div>
+										</ResultList.Description>
+									</ResultList.Content>
+								</ResultList>
+							))}
+						</ReactiveList.ResultListWrapper>
+					)}
+				/>
+			</div>
+		</div>
+	</ReactiveBase>
+);
+
+export default Main;
 
 ReactDOM.render(<Main />, document.getElementById('root'));
