@@ -1,6 +1,11 @@
 /* eslint-disable */
 import React, { Component } from 'react';
-import { ReactiveBase, ToggleButton, SelectedFilters, ResultList } from '@appbaseio/reactivesearch';
+import {
+	ReactiveBase,
+	ToggleButton,
+	SelectedFilters,
+	ReactiveList,
+} from '@appbaseio/reactivesearch';
 
 import initReactivesearch from '@appbaseio/reactivesearch/lib/server';
 
@@ -31,7 +36,13 @@ const resultListProps = {
 	className: 'result-list-container',
 	from: 0,
 	size: 5,
-	onData: ListItemView,
+	render: ({ data }) => (
+		<ReactiveList.ResultListWrapper>
+			{data.map(item => (
+				<ListItemView key={item._id} {...item} />
+			))}
+		</ReactiveList.ResultListWrapper>
+	),
 	pagination: true,
 	react: {
 		and: ['CitySensor'],
@@ -49,7 +60,7 @@ export default class Main extends Component {
 					},
 					{
 						...resultListProps,
-						source: ResultList,
+						source: ReactiveList,
 					},
 				],
 				null,
@@ -68,7 +79,7 @@ export default class Main extends Component {
 						</div>
 						<div className="col">
 							<SelectedFilters />
-							<ResultList {...resultListProps} />
+							<ReactiveList {...resultListProps} />
 						</div>
 					</div>
 				</ReactiveBase>
