@@ -29,6 +29,7 @@ import H2 from '../styles/H2';
 import queries from '../styles/mediaQueries';
 import { getButtonStyle, getLinkStyle } from '../styles/utils';
 import DiscoverRS from './DiscoverRS';
+import AppbaseUsers from './AppbaseUsers';
 
 const button = {
 	fontSize: '14px',
@@ -277,7 +278,8 @@ class HomePage extends Component {
 							<Layout>
 								<div className={titleRow}>
 									{config.banner6.button ? (
-										<H3 style={{
+										<H3
+											style={{
 												paddingBottom: 50,
 											}}
 										>
@@ -375,7 +377,8 @@ class HomePage extends Component {
 										{config.banner6.demos.map((
 											d,
 											index, // eslint-disable-line
-										) => (Object.keys(d).length ? ( // eslint-disable-next-line
+										) =>
+											(Object.keys(d).length ? ( // eslint-disable-next-line
 												<ImageCard key={index} src={d.src}>
 													<div>
 														<Title>{d.title}</Title>
@@ -412,8 +415,9 @@ class HomePage extends Component {
 							<Layout>
 								<div className={titleRow}>
 									{config.banner7.button ? (
-										<H3 style={{
-											paddingBottom: 50,
+										<H3
+											style={{
+												paddingBottom: 50,
 											}}
 										>
 											{config.banner7.title}
@@ -450,10 +454,10 @@ class HomePage extends Component {
 									smGutter="0px"
 									style={{ marginBottom: '50px' }}
 								>
-									{config.banner7.articles.map((d, index) => (
-										Object.keys(d).length ? (
-										// eslint-disable-next-line
-										<ImageCard key={index} src={d.src}>
+									{config.banner7.articles.map((d, index) =>
+										(Object.keys(d).length ? (
+											// eslint-disable-next-line
+											<ImageCard key={index} src={d.src}>
 												<div>
 													<Title>{d.title}</Title>
 													<p>{d.description}</p>
@@ -469,9 +473,10 @@ class HomePage extends Component {
 														Read Now
 													</SecondaryLink>
 												</div>
-										</ImageCard>)
-										: <div />
-									))}
+											</ImageCard>
+										) : (
+											<div />
+										)))}
 								</Grid>
 							</Layout>
 						</Section>
@@ -505,6 +510,7 @@ class HomePage extends Component {
 							<SupportGrid configName={config.name} />
 						</Layout>
 					</Section>
+					<AppbaseUsers />
 					<Footer configName={config.name} footerConfig={config.footer} />
 					<a
 						href={config.producthunt}
